+++ lib/src/trailer.rs
@@ -57,8 +57,8 @@ pub enum TrailerParseError {
 ///
 /// In this case, there are four trailers: two `Co-authored-by` lines, one
 /// `Reviewed-by` line, and one `Change-Id` line.
-pub fn parse_description_trailers(body: &str) -> Vec<Trailer> {
-    let (trailers, blank, found_git_trailer, non_trailer) = parse_trailers_impl(body);
+pub fn parse_description_trailers(body: &str, separators: &str) -> Vec<Trailer> {
+    let (trailers, blank, found_git_trailer, non_trailer) = parse_trailers_impl(body, separators);
     if !blank {
         // no blank found, this means there was a single paragraph, so whatever
         // was found can't come from the trailer
@@ -76,8 +76,8 @@ pub fn parse_description_trailers(body: &str) -> Vec<Trailer> {
 /// Parse the trailers from a trailer paragraph. This function behaves like
 /// `parse_description_trailer`, but will return an error if a blank or
 /// non trailer line is found.
-pub fn parse_trailers(body: &str) -> Result<Vec<Trailer>, TrailerParseError> {
-    let (trailers, blank, _, non_trailer) = parse_trailers_impl(body);
+pub fn parse_trailers(body: &str, separators: &str) -> Result<Vec<Trailer>, TrailerParseError> {
+    let (trailers, blank, _, non_trailer) = parse_trailers_impl(body, separators);
     if blank {
         return Err(TrailerParseError::BlankLine);
     }
@@ -87,14 +87,15 @@ pub fn parse_trailers(body: &str) -> Result<Vec<Trailer>, TrailerParseError> {
     Ok(trailers)
 }
 
-fn parse_trailers_impl(body: &str) -> (Vec<Trailer>, bool, bool, Option<String>) {
+fn parse_trailers_impl(body: &str, separators: &str) -> (Vec<Trailer>, bool, bool, Option<String>) {
     // a trailer always comes at the end of a message; we can split the message
     // by newline, but we need to immediately reverse the order of the lines
     // to ensure we parse the trailer in an unambiguous manner; this avoids cases
     // where a colon in the body of the message is mistaken for a trailer
     let lines = body.trim_ascii_end().lines().rev();
-    let trailer_re =
-        regex::Regex::new(r"^([a-zA-Z0-9-]+) *: *(.*)$").expect("Trailer regex should be valid");
+    let separators_re = regex::escape(separators);
+    let trailer_re = regex::Regex::new(&format!("^([a-zA-Z0-9-]+?) *[{separators_re}] *(.*)$"))
+        .expect("Trailer regex should be valid");
     let mut trailers: Vec<Trailer> = Vec::new();
     let mut multiline_value = vec![];
     let mut found_blank = false;
@@ -157,7 +158,7 @@ mod tests {
             Change-Id: I1234567890abcdef1234567890abcdef12345678
         "#};
 
-        let trailers = parse_description_trailers(descriptions);
+        let trailers = parse_description_trailers(descriptions, ":");
         assert_eq!(trailers.len(), 4);
 
         assert_eq!(trailers[0].key, "Co-authored-by");
@@ -187,7 +188,7 @@ mod tests {
             Change-Id: I1234567890abcdef1234567890abcdef12345678
         "#};
 
-        let trailers = parse_description_trailers(descriptions);
+        let trailers = parse_description_trailers(descriptions, ":");
 
         // should only have Change-Id
         assert_eq!(trailers.len(), 1);
@@ -203,7 +204,7 @@ mod tests {
               newlines in it.
         "#};
 
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
 
         // should only have Change-Id
         assert_eq!(trailers.len(), 1);
@@ -226,14 +227,14 @@ mod tests {
             Signed-off-by: Lucky K Maintainer <lucky@maintainer.example.org>
         "#};
 
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
         assert_eq!(trailers.len(), 2);
     }
 
     #[test]
     fn test_trailers_with_single_line_description() {
         let description = r#"chore: update itertools to version 0.14.0"#;
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
         assert_eq!(trailers.len(), 0);
     }
 
@@ -243,7 +244,7 @@ mod tests {
             foo: 1
             bar: 2
         "#};
-        let res = parse_trailers(trailers_txt);
+        let res = parse_trailers(trailers_txt, ":");
         let trailers = res.expect("trailers to be valid");
         assert_eq!(trailers.len(), 2);
         assert_eq!(trailers[0].key, "foo");
@@ -259,7 +260,7 @@ mod tests {
 
             foo: 2
         "#};
-        let res = parse_trailers(trailers);
+        let res = parse_trailers(trailers, ":");
         assert!(matches!(res, Err(TrailerParseError::BlankLine)));
     }
 
@@ -269,7 +270,7 @@ mod tests {
             bar
             foo: 1
         "#};
-        let res = parse_trailers(trailers);
+        let res = parse_trailers(trailers, ":");
         assert!(matches!(
             res,
             Err(TrailerParseError::NonTrailerLine { line: _ })
@@ -284,7 +285,7 @@ mod tests {
             foo: 1
 
         "#};
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
         assert_eq!(trailers.len(), 1);
     }
 
@@ -297,7 +298,7 @@ mod tests {
 
             bar: 2
         "#};
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
         assert_eq!(trailers.len(), 1);
     }
 
@@ -307,7 +308,7 @@ mod tests {
             subject: whatever
             foo: 1
         "#};
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
         assert_eq!(trailers.len(), 0);
     }
 
@@ -318,7 +319,7 @@ mod tests {
 
              foo: 1
         "#};
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
         assert_eq!(trailers.len(), 0);
     }
 
@@ -329,7 +330,7 @@ mod tests {
 
             foo : 1
         "#};
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
         assert_eq!(trailers.len(), 1);
         assert_eq!(trailers[0].key, "foo");
     }
@@ -341,7 +342,7 @@ mod tests {
 
             foo:  1\x20
         "};
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
         assert_eq!(trailers.len(), 1);
         assert_eq!(trailers[0].value, "1");
     }
@@ -354,7 +355,7 @@ mod tests {
             foo:  1\x20
              2\x20
         "};
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
         assert_eq!(trailers.len(), 1);
         assert_eq!(trailers[0].value, "1 \n 2");
     }
@@ -367,7 +368,7 @@ mod tests {
             foo:  1\x20
             bar:  2\x20
         "};
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
         assert_eq!(trailers.len(), 2);
         assert_eq!(trailers[0].value, "1");
         assert_eq!(trailers[1].value, "2");
@@ -380,7 +381,7 @@ mod tests {
 
             foo:1
         "#};
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
         assert_eq!(trailers.len(), 1);
     }
 
@@ -391,7 +392,7 @@ mod tests {
 
             foo:
         "#};
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
         assert_eq!(trailers.len(), 1);
     }
 
@@ -402,7 +403,7 @@ mod tests {
 
             f_o_o: bar
         "#};
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
         assert_eq!(trailers.len(), 0);
     }
 
@@ -414,7 +415,7 @@ mod tests {
             foo: bar
             baz
         "#};
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
         assert_eq!(trailers.len(), 0);
     }
 
@@ -427,14 +428,14 @@ mod tests {
 
             baz
         "#};
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
         assert_eq!(trailers.len(), 0);
     }
 
     #[test]
     fn test_empty_description() {
         let description = "";
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
         assert_eq!(trailers.len(), 0);
     }
 
@@ -447,9 +448,31 @@ mod tests {
             foo: bar
             (cherry picked from commit 72bb9f9cf4bbb6bbb11da9cda4499c55c44e87b9)
         "#};
-        let trailers = parse_description_trailers(description);
+        let trailers = parse_description_trailers(description, ":");
         assert_eq!(trailers.len(), 1);
         assert_eq!(trailers[0].key, "foo");
         assert_eq!(trailers[0].value, "bar");
+    }
+
+    #[test]
+    fn test_trailer_separators() {
+        let description = indoc! {r#"
+            subject
+
+            BUG-13
+            key|value
+            foo: bar
+            toto > titi
+        "#};
+        let trailers = parse_description_trailers(description, ":-|>");
+        assert_eq!(trailers.len(), 4);
+        assert_eq!(trailers[0].key, "BUG");
+        assert_eq!(trailers[0].value, "13");
+        assert_eq!(trailers[1].key, "key");
+        assert_eq!(trailers[1].value, "value");
+        assert_eq!(trailers[2].key, "foo");
+        assert_eq!(trailers[2].value, "bar");
+        assert_eq!(trailers[3].key, "toto");
+        assert_eq!(trailers[3].value, "titi");
     }
 }