+++ CHANGELOG.md
@@ -355,6 +355,8 @@ Thanks to the people who made this release happen!
   `colors."diff removed token" = { bg = "ansi-color-52", underline = false }`
   will apply a dark red background on removed words in diffs.
 
+* Trailers separators are now specified in the toml configuration.
+
 ### Fixed bugs
 
 * `jj file annotate` can now process files at a hidden revision.